experiment(cli): expose deno cli as a lib to enable embedding
