+++ cli/logger.rs
@@ -43,7 +43,7 @@ impl log::Log for CliLogger {
   }
 }
 
-pub(crate) fn init(maybe_level: Option<log::Level>) {
+pub fn init(maybe_level: Option<log::Level>) {
   let log_level = maybe_level.unwrap_or(log::Level::Info);
   let logger = env_logger::Builder::from_env(
     env_logger::Env::default()
